Suggest categories (#26)
* Adds credential examples and app.json for Heroku

We add app.json, with variables required for Heroku to boot up.

We also add config/application.yml.example for dev setup.

* Titleize suggestions on create, not on update

For formatting purposes we want to assume that category suggestions
should be titleized. If we want to override this we can update it later.

Normally you can do this on mobile with `autocorrect="words"` but theres
a longstanding bug on iOS with this, so we're skipping it.

Closes #23

* Adds Possibly Similar Categories on the backend

Since users will sometime add multiple suggested categories very similar
to each other before we can approve them, we need a way to "link"
suggestions. Now we do.

We add fuzzy searching to categories, allowing us to show possible
options when auditing. Clicking on a link merges it all.

Closes #9