Add exporting (#37)
We need to add exporters. There's a possibility for different formats so
we need need to support multiple formats. This adds an "exporters" class
and folder to do that. Right now the only exporter avaiable on the front
end is the binary, but the support and testing for others is there.

I"ll probably add support for a generator too.