Show instructions page on first login
We want users to see the instructions the first time they log in. This
commit enables the Devise::Trackable module and leverages it to do so.

Note, however, that we do NOT want to track IP address, so we don't add
those columns to the `users` table and override the accessor methods on
the `User` model with no-ops.

Affects #35

Signed-off-by: Justin Reese <justin@justinreese.com>