Rename choose-winnowable-claim method for clarity
When winnowing, we fetch a random unchecked claim and show it to the
user. Previously, this was wrapped up in a private method called
`new_claim`, which is a little confusingly-named. This commit renames
it to the more clear `winnowable_claim`.

(I considered naming it `random_unchecked_claim`, but that binds the
name to the current selection model too tightly; we’ll likely improve
that selection model in the future, e.g. #51.)

Affects #41

Signed-off-by: Justin Reese <justin@justinreese.com>