Gracefully handle a lack of winnowable claims
Before, if there were no winnowable claims to show the user, the app
just crashed.

Now, we redirect to the home with a happy notice.

This commit also moves the `authenticate_user!` action to the top of
the list, so it’s performed first. (Otherwise, we’ll run the other two
actions before authenticating the user.)

Affects #41

Signed-off-by: Justin Reese <justin@justinreese.com>